Update portfolio
- Reorganized files
- Added dark mode support
- Updated resume
