--- src/App.css
@@ -1,65 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.ityped-cursor {
-  font-size: 1rem;
-  opacity: 1;
-  -webkit-animation: blink 0.3s infinite;
-  -moz-animation: blink 0.3s infinite;
-  animation: blink 0.3s infinite;
-  animation-direction: alternate;
-}
-
-@keyframes blink {
-  100% {
-      opacity: 0;
-  }
-}
-
-@-webkit-keyframes blink {
-  100% {
-      opacity: 0;
-  }
-}
-
-@-moz-keyframes blink {
-  100% {
-      opacity: 0;
-  }
-}
