+++ src/index.css
@@ -1,3 +1,15 @@
+@media (prefers-color-scheme: dark) {
+  html {
+    background-color: #303030;
+  }
+}
+
+@media (prefers-color-scheme: light) {
+  html {
+    background-color: #fafafa;
+  }
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
